`ADD` `GITMODULE` maqueta-ong
This project, lovingly crafted by @bairon-cano, @JuniorCalva40, and @CMacxC, is for the "Fundacion Benefica Internacional Vash Gorod". It serves as a template to assist new developers in creating pages for NGOs.

The project has been incorporated as a submodule, given that it maintains its own repository, which can be found here: `https://github.com/inkua/maqueta-ong`.

For further information on git modules, GitHub provides comprehensive documentation that can be accessed here: `https://github.blog/2016-02-01-working-with-submodules/`.

Co-Authored-By: Mario A. <93497141+CMacxC@users.noreply.github.com>
Co-Authored-By: Alejandro Calva <101780546+JuniorCalva40@users.noreply.github.com>
Co-Authored-By: Bairon Cano <84099271+bairon-cano@users.noreply.github.com>